layer: refactor MapOptions to UnpackOptions
A long time ago as a hack to avoid this very patch, I added KeepDirlinks to
MapOptions. Then, I added a callback and the StartFrom argument, having to
thread that through all the callsites. In the next patch, I'll add another
unpack option.

So, let's bite the bullet and create an UnpackOptions struct to pass to the
unpacking code. Note that this doesn't have *all* of the options like
bundle path and such, since those aren't necessarily relevant for all
unpacking operations (UnpackRootfs(), UnpackManifest(), Unpack(), etc.).

Signed-off-by: Tycho Andersen <tycho@tycho.ws>