layer: add support for translating overlay whiteouts to OCI ones
This is the compliment to the previous commit, in case people want to
re-pack layers using the OCI standard instead of the overlayfs standard.

Also add a PackOptions{} struct, similar to UnpackOptions{}, which
can control the low level behavior of the packing routines. I expect we'll
use this later for hooks, too.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>